requirements.txt
contextlib2==0.5.3
dj-database-url==0.4.1
Django==1.9.6
djangorestframework==3.3.3
django-filter==0.13.0
google-api-python-client==1.5.0
gunicorn==19.5.0
httplib2==0.9.2
oauth2client==2.0.2
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-decouple==3.0
raven==5.18.0
requests==2.10.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
whitenoise==3.0
